Only map rows to ordinals once for hybrid vector queries

+++ src/java/org/apache/cassandra/index/sai/disk/vector/OrdinalsView.java
@@ -30,14 +30,15 @@ public interface OrdinalsView extends AutoCloseable
 {
     interface OrdinalConsumer
     {
-        void accept(long rowId, int ordinal) throws IOException;
+        void accept(int rowId, int ordinal) throws IOException;
     }
 
     /** return the vector ordinal associated with the given row, or -1 if no vectors are associated with it */
     int getOrdinalForRowId(int rowId) throws IOException;
 
     /**
-     * iterates over all ordinals in the view.  order of iteration is undefined.
+     * iterates over all ordinals in the view.  order of iteration is undefined. Only calls consumer for valid mappings
+     * from row id to ordinal.
      */
     @VisibleForTesting
     void forEachOrdinalInRange(int startRowId, int endRowId, OrdinalConsumer consumer) throws IOException;
